feat: made changes to sidebar component with erro
