+++ src/App.jsx
@@ -6,7 +6,7 @@ import SideNavigation from './components/global/SideNavigation'
 function App() {
   return (
     <>
-    <SideNavigation/>
+
      <BrowserRouter>
   <Routes>
  <Route path="/" element={<Home />} />
