--- src/components/global/SideNavigation.jsx
@@ -1,21 +1,0 @@
-import { Sidebar, Menu, MenuItem } from 'react-pro-sidebar';
-import { Link } from 'react-router-dom';
-
-
-function SideNavigation() {
-  return (
-    <>
-    <Sidebar>
-  <Menu>
-    <MenuItem routerLink={<Link to="/documentation" />}> Documentation</MenuItem>
-    <MenuItem routerLink={<Link to="/calendar" />}> Calendar</MenuItem>
-    <MenuItem routerLink={<Link to="/e-commerce" />}> E-commerce</MenuItem>
-  </Menu>
-</Sidebar>;
-
-    </>
-
-  )
-}
-
-export default SideNavigation
