--- src/page/hello.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function hello() {
-  return (
-    <div>hello</div>
-  )
-}
-
-export default hello
